requirements.txt
pyuac
python-pptx
pytesseract
python-docx
PyPDF2
pywin32